完成AI合同查询Agent基础版本，准备开发DataProfiler Agent

+++ tools/query_tools.py
@@ -190,9 +190,11 @@ def query_value_tool(state):
 
         "期末余额总额":
 
-            round(
-                result["期末余额"].sum(),
-                2
+            float(
+                round(
+                    result["期末余额"].sum(),
+                    2
+                )
             )
